tweak more Universe dependencies
