+++ Empire/ProductionQueue.cpp
@@ -4,9 +4,11 @@
 #include "../universe/BuildingType.h"
 #include "../universe/Condition.h"
 #include "../universe/Planet.h"
+#include "../universe/ScriptingContext.h"
 #include "../universe/ShipHull.h"
 #include "../universe/ShipPart.h"
 #include "../universe/ShipDesign.h"
+#include "../universe/Universe.h"
 #include "../universe/ValueRef.h"
 #include "../util/AppInterface.h"
 #include "../util/GameRules.h"
