+++ universe/Fighter.h
@@ -3,9 +3,11 @@
 
 
 #include "UniverseObject.h"
-#include "../universe/Condition.h"
 #include "../util/Export.h"
 
+namespace Condition {
+    struct Condition;
+}
 
 ////////////////////////////////////////////////
 // Fighter
